Factory & seeder completed, UserController index, show, store

+++ database/migrations/2019_09_13_011927_create_category_product_table.php
@@ -15,8 +15,8 @@ class CreateCategoryProductTable extends Migration
     {
         Schema::create('category_product', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('category_id')->unsigned();            
-            $table->integer('product_id')->unsigned();            
+            $table->bigInteger('category_id')->unsigned();            
+            $table->bigInteger('product_id')->unsigned();            
             $table->timestamps();
 
             $table->foreign('category_id')->references('id')->on('categories');
